+++ R/DataClasses.R
@@ -1055,10 +1055,9 @@ setClass("MassifquantParam",
 #' mp
 #'
 #' ## Loading a small subset of direct injection, single spectrum files
-#' library(msdata)
-#' fticrf <- list.files(system.file("fticr-mzML", package = "msdata"),
-#'                     recursive = TRUE, full.names = TRUE)
-#' fticr <- readMSData(fticrf[1], msLevel. = 1, mode = "onDisk")
+#' library(MsDataHub)
+#' fl <- MsDataHub::HAM004_641fE_14.11.07..Exp1.extracted.mzML()
+#' fticr <- readMSData(fl, msLevel. = 1, mode = "onDisk")
 #'
 #' ## Perform the MSW peak detection on these:
 #' p <- MSWParam(scales = c(1, 7), peakThr = 80000, ampTh = 0.005,
@@ -1239,7 +1238,8 @@ setClass("PeakDensityParam",
                    minSamples = "numeric",
                    binSize = "numeric",
                    maxFeatures = "numeric",
-                   ppm = "numeric"),
+                   ppm = "numeric",
+                   rtCenterFun = "character"),
          contains = "Param",
          prototype = prototype(
              sampleGroups = numeric(),
@@ -1248,7 +1248,8 @@ setClass("PeakDensityParam",
              minSamples = 1,
              binSize = 0.25,
              ppm = 0,
-             maxFeatures = 50),
+             maxFeatures = 50,
+             rtCenterFun = "median"),
          validity = function(object) {
              msg <- character()
              if (length(object@bw) > 1 | any(object@bw < 0))