+++ R/methods-MChromatograms.R
@@ -6,11 +6,12 @@
 #'
 #' library(MsExperiment)
 #' library(xcms)
+#' library(faahKO)
 #' ## Perform peak detection on an MChromatograms object
 #'
-#' fls <- c(ko15.CDF(),
-#'          ko16.CDF(),
-#'          ko18.CDF())
+#' fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
+#'          system.file('cdf/KO/ko16.CDF', package = "faahKO"),
+#'          system.file('cdf/KO/ko18.CDF', package = "faahKO"))
 #' od3 <- readMsExperiment(fls)
 #'
 #' ## Disable parallel processing for this example