+++ R/methods-xcmsRaw.R
@@ -2047,6 +2047,8 @@ setMethod("stitch.netCDF.new", "xcmsRaw", function(object, lockMass) {
 #'
 #' @examples
 #' ## Load a test file
+#' library(xcms)
+#' library(MsDataHub)
 #' file <- ko15.CDF()
 #' xraw <- xcmsRaw(file, profstep = 0)
 #' ## The number of scans/spectra:
@@ -2223,6 +2225,8 @@ setMethod("[", signature(x = "xcmsRaw",
 #' @author Johannes Rainer
 #'
 #' @examples
+#' library(xcms)
+#' library(MsDataHub)
 #' file <- ko15.CDF()
 #' ## Load the data without generating the profile matrix (profstep = 0)
 #' xraw <- xcmsRaw(file, profstep = 0)