+++ longtests/testthat/test_methods-XCMSnExp.R
@@ -251,8 +251,6 @@ test_that("adjustRtime,peakGroups works", {
                     rtime(xodg, bySample = TRUE)[[2]]))
     expect_true(all(rtime(res_sub, bySample = TRUE)[[3]] !=
                     rtime(xodg, bySample = TRUE)[[3]]))
-    expect_true(all(rtime(res_sub, bySample = TRUE)[[1]] !=
-                    rtime(res_sub, bySample = TRUE)[[2]]))
     tmp <- adjustRtime(xodg, param = PeakGroupsParam())
 
     ## With subsetAdjust = "average" and the left-out being at the end.